sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "piv-oac"
version = "0.1.0"
description = "SDK Python para el framework PIV/OAC — agentes, contratos y engram store"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.scripts]
piv = "piv_oac.cli.main:cli"

[project.optional-dependencies]
openai = [
    "openai>=1.30.0",
]
telemetry = [
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.10",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["piv_oac*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["piv_oac"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
disallow_untyped_defs = true
warn_return_any = false
warn_unused_ignores = true
exclude = ["tests/"]
